+++ hydra/_internal/utils.py
@@ -4,12 +4,14 @@ import inspect
 import logging.config
 import os
 import sys
+import warnings
 from dataclasses import dataclass
 from os.path import dirname, join, normpath, realpath
 from traceback import print_exc, print_exception
 from types import FrameType
-from typing import Any, Callable, List, Optional, Sequence, Tuple, Union
+from typing import TYPE_CHECKING, Any, Callable, List, Optional, Sequence, Tuple, Union
 
+from omegaconf import DictConfig
 from omegaconf.errors import OmegaConfBaseException
 
 from hydra._internal.config_search_path_impl import ConfigSearchPathImpl
@@ -22,6 +24,9 @@ from hydra.errors import (
     SearchPathException,
 )
 from hydra.types import TaskFunction
+
+if TYPE_CHECKING:
+    from hydra.core.utils import JobReturn
 
 log = logging.getLogger(__name__)
 
@@ -597,3 +602,50 @@ def _get_cls_name(config: Any, pop: bool = True) -> str:
     if not isinstance(classname, str):
         raise InstantiationException("_target_ field type must be a string")
     return classname
+
+
+class Callbacks:
+    def __init__(self, config: DictConfig) -> None:
+        self.callbacks = []
+        from hydra.utils import instantiate
+
+        for params in config.hydra.callbacks.values():
+            self.callbacks.append(instantiate(params))
+
+    def _notify(self, function_name: str, reverse: bool = False, **kwargs: Any) -> None:
+        callbacks = reversed(self.callbacks) if reverse else self.callbacks
+        for c in callbacks:
+            try:
+                getattr(c, function_name)(**kwargs)
+            except Exception as e:
+                warnings.warn(
+                    f"Callback {type(c).__name__}.{function_name} raised {type(e).__name__}: {e}"
+                )
+
+    def on_run_start(self, config: DictConfig, **kwargs: Any) -> None:
+        self._notify(function_name="on_run_start", config=config, **kwargs)
+
+    def on_run_end(self, config: DictConfig, **kwargs: Any) -> None:
+        self._notify(function_name="on_run_end", config=config, reverse=True, **kwargs)
+
+    def on_multirun_start(self, config: DictConfig, **kwargs: Any) -> None:
+        self._notify(function_name="on_multirun_start", config=config, **kwargs)
+
+    def on_multirun_end(self, config: DictConfig, **kwargs: Any) -> None:
+        self._notify(
+            function_name="on_multirun_end", reverse=True, config=config, **kwargs
+        )
+
+    def on_job_start(self, config: DictConfig, **kwargs: Any) -> None:
+        self._notify(function_name="on_job_start", config=config, **kwargs)
+
+    def on_job_end(
+        self, config: DictConfig, job_return: "JobReturn", **kwargs: Any
+    ) -> None:
+        self._notify(
+            function_name="on_job_end",
+            config=config,
+            job_return=job_return,
+            reverse=True,
+            **kwargs,
+        )
